requirements.txt
google-generativeai
flask
flask-cors
SpeechRecognition
pyaudio
pyttsx3
psutil
requests